test: enable internal pipeline job e2e tests
